requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
MouseInfo==0.1.3
pillow==11.0.0
psutil==6.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
urllib3==2.2.3
